Added helper functions to list files, folders, models, phones, and scene types.
The user first needs to import the EXIF data from the csv file as a Pandas data frame. Then these new helper functions take the data frame as input. These functions all assume that the data frame has a column named "SourceFile" that gives the file path and file name of the images.

I haven't written tests for get_files or get_folders yet because the expected results will need to be stored as fixtures.