Added optional encoding argument to read_exif()
The read_exif() function worked fine on the Samsung EXIF data csv files, but returned the error "UnicodeDecodeError: 'charmap' codec can't decode byte 0x81 in position 6792: character maps to <undefined>" on the iPhone EXIF data csv files. One suggestion on StackOverflow is to specify a different encoding in the pd.read_csv function. This is the change that I made.